WebService/user/src/main/java/edu/uoc/epcsd/user/services/UserDetailsServiceImpl.java
package edu.uoc.epcsd.user.services;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import edu.uoc.epcsd.user.repositories.UserRepository;


@Service
public class UserDetailsServiceImpl implements UserDetailsService {

    @Autowired
    UserRepository userRepository;

    @Override
    @Transactional
    public UserDetails loadUserByUsername(String email){
    	return userRepository.findByEmail(email)
                .orElseThrow(
                    () -> new UsernameNotFoundException("[ERROR] User not found")
                    );
    }
}
